# doc/architecture/openapi/iam-v1.yaml
openapi: 3.1.0
info:
  title: Bone IAM API
  version: 1.1.0
  description: |
    身份与访问管理（As-Is，**契约真源**）。规范见 Bone-API-规范.md §13.1；
    详设：doc/design/modules/6. IAM账号权限管理模块详细设计方案.md（§5 与本 yaml 单源约束）。

    认证：JWT Bearer（claim `scopes` = 权限码列表）；管理 API 经 `@EnableMethodSecurity`
    + `@PreAuthorize("hasAuthority('iam:…')")` 校验细粒度权限码。
    租户：`tenantId` 仅来自 JWT claim，写入 `TenantContext`，禁止 Body 覆盖。

    网关前缀：`/api/v1/iam` → bone-iam :8081。
servers:
  - url: /api/v1/iam
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /roles/{id}:
    get:
      summary: 角色详情（含权限摘要）
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
    put:
      summary: 更新角色
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
    delete:
      summary: 删除角色
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
  /roles/{id}/permissions:
    get:
      summary: 查询角色权限
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
    post:
      summary: 分配角色权限
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
...
  /tenants:
    get:
      summary: 租户列表（平台超管）
      description: 权限码 `iam:tenants:read`（仅平台超管角色拥有）。
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: size, in: query, schema: { type: integer, default: 10 } }
      responses:
        '200': { description: OK }
        '403': { description: 缺少 iam:tenants:read }
    post:
      summary: 创建租户
      description: 权限码 `iam:tenants:write`。
      responses:
        '200': { description: OK }
  /tenants/{id}:
    get:
      summary: 租户详情
      description: 权限码 `iam:tenants:read`。
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
    put:
      summary: 更新租户
      description: 权限码 `iam:tenants:write`。
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
    delete:
      summary: 删除租户（级联清理 IAM 子表 + 软删租户）
      description: 权限码 `iam:tenants:write`；禁止删除平台租户 id=0。
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
        '400': { description: IAM_TENANT_DELETE_FORBIDDEN（平台租户） }
  /tenants/{id}/enable:
    post:
      summary: 启用租户
      description: 权限码 `iam:tenants:write`。
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
  /tenants/{id}/disable:
    post:
      summary: 禁用租户
      description: 权限码 `iam:tenants:write`。
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
  /tenants/{id}/quota:
    put:
      summary: 更新租户配额（maxAccounts / maxRoles）
      description: 权限码 `iam:tenants:write`；`null` 表示不限制。
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
        '400': { description: 配额非法 }